refactor: centralizar procesos de email
