+++ src/DALC/ordenes.dalc.ts
@@ -19,7 +19,7 @@ import { Lote } from "../entities/Lote"
 import { ordenEstadoHistorico_insert_DALC, ordenEstadoHistorico_getByIdOrden_DALC } from "./ordenEstadoHistorico.dalc"
 import { OrdenEstadoHistorico } from "../entities/OrdenEstadoHistorico"
 import { emailProcesoConfig_get } from "./emailProcesoConfig.dalc"
-import { EMAIL_PROCESOS } from "../constants/emailProcesos"
+import { EMAIL_PROCESOS } from "../constants/procesosEmail"
 const { logger } = require('../helpers/logger')
 
 // Move orden_getById_DALC to the top to resolve declaration order issue
@@ -49,7 +49,7 @@ export const orden_informarEmisionEtiqueta = async (orden: Orden) => {
         cuerpo += `<br><br><hr>Este mail ha sido enviado por un sistema automatizado e inatendido.  Por favor, no responder.`
 
         const config = await emailProcesoConfig_get(orden.IdEmpresa, EMAIL_PROCESOS.ORDEN_ETIQUETA)
-        const plantilla = await renderEmailTemplate('ORDEN_ETIQUETA', valores, config?.IdEmailTemplate)
+        const plantilla = await renderEmailTemplate(EMAIL_PROCESOS.ORDEN_ETIQUETA, valores, config?.IdEmailTemplate)
         if (plantilla) {
             titulo = plantilla.asunto
             cuerpo = plantilla.cuerpo
